+++ test-e2e/webui.ts
@@ -37,7 +37,8 @@ const common = {
     ],
     formatOptions: {
         console: {
-            showLogs: true
+            showLogs: true,
+            showProgress: false
         }
     },
     memory: new Memory(),